test(memory): split reset recall coverage
